pyproject.toml
[tool.poetry]
name = "cripto-price"
version = "0.1.0"
description = "Crypto Deribit Price Service"
authors = ["Nadezhda_Popova <nadezhdapopova13@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
poetry-core = ">=2.0.0"
fastapi = "^0.135.1"
uvicorn = "^0.41.0"
sqlalchemy = "^2.0.48"
asyncpg = "^0.31.0"
psycopg2-binary = "^2.9.11"
aiohttp = "^3.13.3"
celery = "^5.6.2"
redis = "^7.3.0"
pydantic = "^2.12.5"
python-dotenv = "^1.2.2"
pre-commit = "^4.5.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
lint = [
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]

[tool.black]
line-length = 119
exclude = """
/(
    .git
    .venv
    venv
    __pycache__
)/
"""

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
skip_gitignore = true
